[DEV-1890] fix amlCheck reset fee bug

+++ src/subdomains/core/buy-crypto/process/services/buy-crypto.service.ts
@@ -253,7 +253,7 @@ export class BuyCryptoService {
     const fee = entity.fee;
 
     await this.buyCryptoRepo.update(...entity.resetAmlCheck());
-    await this.buyCryptoRepo.deleteFee(fee);
+    if (fee) await this.buyCryptoRepo.deleteFee(fee);
   }
 
   async getUserTransactions(
